# Conflicts:
#	controllers/client_article.py

+++ controllers/admin_client.py
@@ -18,19 +18,53 @@ def admin_gestion():
             WHERE USER.id_user != %s;'''
     mycursor.execute(sql, insert)
     user = mycursor.fetchall()
-    print(user)
     return render_template('admin/client/gestion_client.html', user=user)
 
 @admin_client.route('/admin/client/gestion/delete', methods=['GET'])
 def delete_client():
     mycursor = get_db().cursor()
     id_client = request.args.get('id')
-    print("Je delete le client id : "+id_client)
     tuple_insert = id_client
-    sql = '''DELETE FROM USER WHERE USER.id_user = %s;'''
+    sql = '''SELECT * FROM COMMANDE 
+            INNER JOIN user u on commande.id_user = u.id_user
+            INNER JOIN etat e on commande.id_etat = e.id_etat
+            WHERE COMMANDE.id_user = %s;'''
+    mycursor.execute(sql, tuple_insert)
+    commande = mycursor.fetchall()
+    print(commande)
+    if commande:
+        return render_template('admin/client/delete_client.html', commande=commande)
+    else:
+        sql = '''DELETE FROM USER WHERE USER.id_user = %s;'''
+        mycursor.execute(sql, tuple_insert)
+        get_db().commit()
+    print("Je delete le client id : "+id_client)
+    return redirect("/admin/client/gestion")
+
+@admin_client.route('/admin/client/gestion/delete/commande', methods=['GET'])
+def delete_client_user():
+    mycursor = get_db().cursor()
+    id_commande = request.args.get('id')[0]
+    id_client = request.args.get('id')[1]
+    tuple_insert = id_commande
+    sql = '''DELETE FROM COMMANDE WHERE COMMANDE.id_commande = %s;'''
     mycursor.execute(sql, tuple_insert)
     get_db().commit()
-    return redirect("/admin/client/gestion")
+    tuple_insert = id_client
+    sql = '''SELECT * FROM COMMANDE 
+            INNER JOIN user u on commande.id_user = u.id_user
+            INNER JOIN etat e on commande.id_etat = e.id_etat
+            WHERE COMMANDE.id_user = %s;'''
+    mycursor.execute(sql, tuple_insert)
+    commande = mycursor.fetchall()
+    if commande != ():
+        return render_template('admin/client/delete_client.html', commande=commande)
+    else:
+        sql = '''DELETE FROM USER WHERE USER.id_user = %s;'''
+        mycursor.execute(sql, tuple_insert)
+        get_db().commit()
+        return redirect('/admin/client/gestion')
+
 
 @admin_client.route('/admin/client/gestion/add', methods=['GET'])
 def add_client():
@@ -58,11 +92,32 @@ def valid_add_client():
     tuple_insert = (username, email, password, role, adresse)
     sql = '''INSERT INTO user(username_user,email_user,password_user,role_user, adresse_user) VALUES (%s,%s,%s,%s,%s);'''
     mycursor.execute(sql, tuple_insert)
-    get_db().commit()  # position de cette ligne discutatble !
-    sql = '''SELECT last_insert_id() AS last_insert_id;'''
-    mycursor.execute(sql)
-    info_last_id = mycursor.fetchone()
-    user_id = info_last_id['last_insert_id']
-    print('last_insert_id', user_id)
+    get_db().commit()
+    print("J'ajoute un utilisateur")
+    return redirect('/admin/client/gestion')
+
+@admin_client.route('/admin/client/gestion/edit/<int:id>', methods=['GET'])
+def edit_client(id):
+    mycursor = get_db().cursor()
+    tuple_insert = id
+    query = '''SELECT * FROM USER WHERE USER.id_user = %s ;'''
+    mycursor.execute(query, tuple_insert)
+    user = mycursor.fetchone()
+    return render_template('admin/client/edit_client.html', user=user)
+
+
+@admin_client.route('/admin/client/gestion/edit', methods=['POST'])
+def valid_edit_client():
+    mycursor = get_db().cursor()
+    id_user = request.form.get('id_user')
+    print("Je modifie l'utilisateur : " + id_user)
+    username_user = request.form.get('username_user')
+    role_user = request.form.get('role_user')
+    email_user = request.form.get('email_user')
+    adresse_user = request.form.get('adresse_user')
+    tuple_insert = (username_user, role_user, email_user, adresse_user, id_user)
+    query = '''UPDATE USER SET username_user = %s, role_user = %s, email_user = %s, adresse_user = %s
+                WHERE id_user = %s ;'''
+    mycursor.execute(query, tuple_insert)
     get_db().commit()
     return redirect('/admin/client/gestion')
